+++ ipv6service/api/src/main/java/org/opendaylight/netvirt/ipv6service/api/IVirtualNetwork.java
@@ -7,16 +7,16 @@
  */
 package org.opendaylight.netvirt.ipv6service.api;
 
-import java.math.BigInteger;
 import java.util.List;
 import org.opendaylight.yang.gen.v1.urn.ietf.params.xml.ns.yang.ietf.yang.types.rev130715.Uuid;
+import org.opendaylight.yangtools.yang.common.Uint64;
 
 public interface IVirtualNetwork {
     Uuid getNetworkUuid();
 
     Long getElanTag();
 
-    List<BigInteger> getDpnsHostingNetwork();
+    List<Uint64> getDpnsHostingNetwork();
 
     int getMtu();
 }
