+++ ipv6service/api/src/main/java/org/opendaylight/netvirt/ipv6service/api/IVirtualPort.java
@@ -7,10 +7,10 @@
  */
 package org.opendaylight.netvirt.ipv6service.api;
 
-import java.math.BigInteger;
 import java.util.List;
 import org.opendaylight.yang.gen.v1.urn.ietf.params.xml.ns.yang.ietf.inet.types.rev130715.Ipv6Address;
 import org.opendaylight.yang.gen.v1.urn.ietf.params.xml.ns.yang.ietf.yang.types.rev130715.Uuid;
+import org.opendaylight.yangtools.yang.common.Uint64;
 
 public interface IVirtualPort {
 
@@ -22,7 +22,7 @@ public interface IVirtualPort {
 
     String getDeviceOwner();
 
-    BigInteger getDpId();
+    Uint64 getDpId();
 
     List<Ipv6Address> getIpv6Addresses();
 }
