+++ natservice/impl/src/main/java/org/opendaylight/netvirt/natservice/internal/FloatingIPHandler.java
@@ -7,23 +7,23 @@
  */
 package org.opendaylight.netvirt.natservice.internal;
 
-import java.math.BigInteger;
-
 import org.opendaylight.genius.infra.Datastore.Configuration;
 import org.opendaylight.genius.infra.TypedReadWriteTransaction;
 import org.opendaylight.yang.gen.v1.urn.ietf.params.xml.ns.yang.ietf.yang.types.rev130715.Uuid;
 import org.opendaylight.yang.gen.v1.urn.opendaylight.netvirt.natservice.rev160111.ProviderTypes;
 import org.opendaylight.yang.gen.v1.urn.opendaylight.netvirt.natservice.rev160111.floating.ip.info.router.ports.ports.InternalToExternalPortMap;
+import org.opendaylight.yangtools.yang.common.Uint32;
+import org.opendaylight.yangtools.yang.common.Uint64;
 
 public interface FloatingIPHandler {
 
-    void onAddFloatingIp(BigInteger dpnId, String routerUuid, long routerId, Uuid networkId, String interfaceName,
+    void onAddFloatingIp(Uint64 dpnId, String routerUuid, Uint32 routerId, Uuid networkId, String interfaceName,
                          InternalToExternalPortMap mapping, TypedReadWriteTransaction<Configuration> confTx);
 
-    void onRemoveFloatingIp(BigInteger dpnId, String routerUuid, long routerId, Uuid networkId,
-                            InternalToExternalPortMap mapping, long label,
+    void onRemoveFloatingIp(Uint64 dpnId, String routerUuid, Uint32 routerId, Uuid networkId,
+                            InternalToExternalPortMap mapping, Uint32 label,
                             TypedReadWriteTransaction<Configuration> confTx);
 
-    void cleanupFibEntries(BigInteger dpnId, String vpnName, String externalIp, long label,
+    void cleanupFibEntries(Uint64 dpnId, String vpnName, String externalIp, Uint32 label,
                            TypedReadWriteTransaction<Configuration> confTx, ProviderTypes provType);
 }
